requirements.txt
# Core Flask stack
Flask==2.3.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
requests==2.31.0

# OAuth and Google API
Authlib==1.2.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# PostgreSQL
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0

# Domain checker 
dnspython==2.4.2
schedule==1.2.0